[IssueType] link issue type to user

+++ src/issue-type/issue-type.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { from } from 'rxjs';
 import { IssueType, IssueTypeDocument } from './issue-type.schema';
+import { switchMap } from 'rxjs/operators';
 
 
 @Injectable()
@@ -12,27 +13,54 @@ export class IssueTypeService {
   ) {
   }
 
-  listAction() {
+  listAction(userId: string) {
     return from (
-      this.model.find()
+      this.model.find({ownedBy: userId})
     )
   }
 
-  createAction(issueType: IssueType) {
+  createAction(issueType: IssueType, userId: string) {
     return from(
-      new this.model(issueType).save()
+      new this.model({
+        name: issueType.name,
+        ownedBy: userId
+      }).save()
     )
   }
 
-  updateAction(id: string, {name}) {
+  updateAction(id: string, userId: string, {name}) {
     return from(
-      this.model.findOneAndUpdate({_id: id}, {name})
-    )
+      this.model.findOne({_id: id, ownedBy: userId})
+    ).pipe(
+      switchMap(issueType => {
+        if (!issueType) {
+          throw new NotFoundException(
+            `Issue type with ID: ${id} does not exist or does not belong to you`,
+          );
+        }
+
+        return from(
+          this.model.findOneAndUpdate({_id: id, ownedBy: userId}, {name})
+        )
+      })
+    );
   }
 
-  deleteAction(id: string) {
+  deleteAction(id: string, userId: string) {
     return from(
-      this.model.findByIdAndDelete({_id: id})
-    )
+      this.model.findOne({_id: id, ownedBy: userId})
+    ).pipe(
+      switchMap(issueType => {
+        if (!issueType) {
+          throw new NotFoundException(
+            `Issue type with ID: ${id} does not exist or does not belong to you`
+          );
+        }
+
+        return from(
+          this.model.findByIdAndDelete({_id: id, ownedBy: userId})
+        )
+      })
+    );
   }
 }
